requirements.txt
python-telegram-bot~=21.6
logging~=0.4.9.6
python-dotenv~=1.2.1